# web/assets/openapi.yaml
openapi: 3.0.3
info:
  title: RenoP HTTP API
  version: "2.0"
  contact:
    url: https://github.com/404Setup/RenoP
  description: |
    Management and registry APIs for RenoP, a self-hosted artifact repository server
    supporting Maven, unstructured file storage, npm, Cargo (Rust sparse index), and Docker/OCI registries.

    Default listen address: `0.0.0.0:3000`.

    **Route layout:**
    - Management API: `/api/...`
    - Maven artifact storage: `/{repo}/{path}` (Maven 2 layout)
    - Unstructured file storage: `/{repo}/{path}`
    - Docker/OCI Distribution API: `/v2/...`
    - Cargo sparse registry: `/{repo}/config.json` and `/{repo}/api/v1/...`
    - npm registry: `/{repo}/{package}` and `/{repo}/-/...`
    - Javadoc preview: `/javadoc/{repo}/{path}`
    - Cargo-doc preview: `/cargodoc/{repo}/{path}`

    **Wire format:**
    Most management endpoints use `application/x-protobuf`. Schemas here describe the
    logical fields (snake_case) for documentation and client stub generation.
    See `proto/api/v1/api.proto` for canonical protobuf definitions.
    Docker/OCI, npm, and Cargo endpoints use JSON.

    **Error bodies:**
    Management errors are plain text (`Unauthorized`, `Forbidden`, `Not found`).
    Docker browser-management errors additionally expose a stable `X-Renop-Error-Code` header for localization.
    Docker Distribution API errors follow `{"errors":[{"code":"...","message":"...","detail":"..."}]}`.
...

  license:
    name: MPL-2.0
...
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: renop_session
      description: HttpOnly session cookie issued by `POST /api/auth/login`. Idle timeout is approximately 7 days.
    basicAuth:
      type: http
      scheme: basic
      description: |
        `username:password` or `username:API-token` encoded as HTTP Basic credentials.
        Basic credentials are accepted only by standard package protocols, Docker token exchange, and chunked upload.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A prefixed API-token secret. API endpoints intersect the token's scopes with the owning account's current
        permissions. Query-string credentials and `Authorization: Session` are rejected.
  parameters:
...
    StatisticsGroupBy:
      name: group_by
      in: query
      schema:
        type: string
        enum: [ user, repository, namespace, package, version ]
        default: repository
    StatisticsRepository:
      name: repository
      in: query
      schema:
        type: string
    StatisticsFormat:
      name: format
      in: query
      schema:
        type: string
        enum: [ maven, files, cargo, docker, npm ]
    StatisticsNamespace:
      name: namespace
      in: query
      schema:
        type: string
    StatisticsPackage:
      name: package
      in: query
      schema:
        type: string
    StatisticsVersion:
      name: version
      in: query
      schema:
        type: string
    StatisticsLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    StatisticsOffset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
        default: 0
...
  schemas:
...
    DownloadStatisticRecord:
      type: object
      required: [ count, bytes, updated_at ]
      properties:
        username:
          type: string
        repository:
          type: string
        format:
          type: string
          enum: [ maven, files, cargo, docker, npm ]
        namespace:
          type: string
          description: Verified Maven publishing domain or npm scope when applicable.
        package:
          type: string
          description: Maven coordinate, npm package, Cargo crate, Docker image, or unstructured file path.
        version:
          type: string
        count:
          type: integer
          format: int64
          minimum: 0
        bytes:
          type: integer
          format: int64
          minimum: 0
        updated_at:
          type: integer
          format: int64
    DownloadStatisticsPage:
      type: object
      required: [ group_by, count, bytes, records, total, limit, offset ]
      properties:
        group_by:
          type: string
          enum: [ user, repository, namespace, package, version ]
        count:
          type: integer
          format: int64
        bytes:
          type: integer
          format: int64
        records:
          type: array
          items:
            $ref: "#/components/schemas/DownloadStatisticRecord"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
paths:
...
  /api/statistics:
    get:
      tags: [ statistics ]
      summary: Query the API-token owner's download statistics
      description: Requires a Bearer API token with `statistics:read`; browser sessions and Basic credentials are rejected.
      operationId: getOwnDownloadStatistics
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: "#/components/parameters/StatisticsGroupBy"
        - $ref: "#/components/parameters/StatisticsRepository"
        - $ref: "#/components/parameters/StatisticsFormat"
        - $ref: "#/components/parameters/StatisticsNamespace"
        - $ref: "#/components/parameters/StatisticsPackage"
        - $ref: "#/components/parameters/StatisticsVersion"
        - $ref: "#/components/parameters/StatisticsLimit"
        - $ref: "#/components/parameters/StatisticsOffset"
      responses:
        "200":
          description: Statistics grouped at the requested level.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadStatisticsPage"
  /api/statistics/users/{username}:
    get:
      tags: [ statistics ]
      summary: Query one user's download statistics
      description: The token owner may query itself. Another account requires a system administrator token with `admin:statistics`.
      operationId: getUserDownloadStatistics
      security:
        - bearerAuth: [ ]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/StatisticsGroupBy"
        - $ref: "#/components/parameters/StatisticsRepository"
        - $ref: "#/components/parameters/StatisticsFormat"
        - $ref: "#/components/parameters/StatisticsNamespace"
        - $ref: "#/components/parameters/StatisticsPackage"
        - $ref: "#/components/parameters/StatisticsVersion"
        - $ref: "#/components/parameters/StatisticsLimit"
        - $ref: "#/components/parameters/StatisticsOffset"
      responses:
        "200":
          description: User-bound statistics page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadStatisticsPage"
  /api/statistics/system:
    get:
      tags: [ statistics ]
      summary: Query system-wide download statistics
      description: Requires a system administrator Bearer API token with `admin:statistics`.
      operationId: getSystemDownloadStatistics
      security:
        - bearerAuth: [ ]
      parameters:
        - name: username
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/StatisticsGroupBy"
        - $ref: "#/components/parameters/StatisticsRepository"
        - $ref: "#/components/parameters/StatisticsFormat"
        - $ref: "#/components/parameters/StatisticsNamespace"
        - $ref: "#/components/parameters/StatisticsPackage"
        - $ref: "#/components/parameters/StatisticsVersion"
        - $ref: "#/components/parameters/StatisticsLimit"
        - $ref: "#/components/parameters/StatisticsOffset"
      responses:
        "200":
          description: System-wide statistics page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadStatisticsPage"

  # ========== Updater ==========
...